add view score

+++ app/Http/Controllers/admin/DanhGiaController.php
@@ -12,21 +12,33 @@ class DanhGiaController extends Controller
 {
     public function index()
     {
-        $khoahoc = KhoaHoc::paginate(10);
-        return view('admin.danhgia.index', ['khoahoc' => $khoahoc]);
-    } 
+        if (session('login')->LOAITK == 3) {
+            $khoahoc = KhoaHoc::paginate(10);
+            return view('admin.danhgia.index', ['khoahoc' => $khoahoc]);
+        } else {
+            return view('admin.thongbao.index');
+        }
+    }
     public function viewdanhgia($id)
     {
-        $danhgia = DanhGia::where('MAKH', '=', $id)->paginate(10);
-        $khoahoc = KhoaHoc::find($id);
-        return view('admin.danhgia.ctdanhgia' , compact('danhgia' , 'id' , 'khoahoc'));
-    }     
+        if (session('login')->LOAITK == 3) {
+            $danhgia = DanhGia::where('MAKH', '=', $id)->paginate(10);
+            $khoahoc = KhoaHoc::find($id);
+            return view('admin.danhgia.ctdanhgia', compact('danhgia', 'id', 'khoahoc'));
+        } else {
+            return view('admin.thongbao.index');
+        }
+    }
 
     public function delete($id)
     {
-        $danhgia = DanhGia::find($id);
-        $danhgia->delete();
+        if (session('login')->LOAITK == 3) {
+            $danhgia = DanhGia::find($id);
+            $danhgia->delete();
 
-        return redirect('admin/danhgia/ctdanhgia/{{$id}}')->with('thongbao', 'Xóa thành công!');
+            return redirect('admin/danhgia/ctdanhgia/{{$id}}')->with('thongbao', 'Xóa thành công!');
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
 }
