+++ app/Http/Controllers/admin/KhuyenMaiController.php
@@ -14,65 +14,68 @@ class KhuyenMaiController extends Controller
 {
     public function index()
     {
-        $today = date("Y-m-d h:i:sa");
+        if (session('login')->LOAITK == 3) {
+            $today = date("Y-m-d H:i:s");
 
-        KhuyenMai::where('NGAYBD', '>',  $today)
-        ->update(['MATT' => '1']);
+            KhuyenMai::where('NGAYBD', '>',  $today)
+                ->update(['MATT' => '1']);
 
-        KhuyenMai::where('NGAYBD', '<=',  $today)
-        ->where('NGAYKT', '=>', $today)
-        ->update(['MATT' => '2']);
+            KhuyenMai::where('NGAYBD', '<=',  $today)
+                ->where('NGAYKT', '=>', $today)
+                ->update(['MATT' => '2']);
 
-        KhuyenMai::where('NGAYKT', '<', $today)
-        ->update(['MATT' => '3']);
+            KhuyenMai::where('NGAYKT', '<', $today)
+                ->update(['MATT' => '3']);
 
-        $khuyenmai = KhuyenMai::paginate(10);      
-        return view('admin.khuyenmai.index', ['khuyenmai'=>$khuyenmai]);
+            $khuyenmai = KhuyenMai::paginate(10);
+            return view('admin.khuyenmai.index', ['khuyenmai' => $khuyenmai]);
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
-    
+
     public function create()
     {
-        $khoahoc = KhoaHoc::all();
-        return view('admin.khuyenmai.create', ['khoahoc'=>$khoahoc]);
+        if (session('login')->LoaiTK == 3) {
+            $khoahoc = KhoaHoc::all();
+            return view('admin.khuyenmai.create', ['khoahoc' => $khoahoc]);
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
 
     public function store(Request $request)
-    {        
-        $this->validate(
-            $request,
-            [
-                'TENKM' => 'unique:KhuyenMai,TENKM|min:10|max:100'
-            ],
-            [
-                'TENKM.unique' => 'Tên khuyến mãi đã tồn tại',
-                'TENKM.min' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự',
-                'TENKM.max' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự'
-            ]
-        );
+    {
+        if (session('login')->LoaiTK == 3) {
+            $this->validate(
+                $request,
+                [
+                    'TENKM' => 'unique:KhuyenMai,TENKM|min:10|max:100'
+                ],
+                [
+                    'TENKM.unique' => 'Tên khuyến mãi đã tồn tại',
+                    'TENKM.min' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự',
+                    'TENKM.max' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự'
+                ]
+            );
 
-        $data = $request->all();   
-        $dateBD = Carbon::createFromTimestamp(strtotime($data['NGAYBD'] . $data['TDBD']));      
-        $dateKT = Carbon::createFromTimestamp(strtotime($data['NGAYKT'] . $data['TDKT']));   
-        $today = date("Y-m-d h:i:sa");
-            if(strtotime($dateBD) >  strtotime($today))
-            {
-                $MATT = 1;
-            }    
-            else if(strtotime($dateBD) <=  strtotime($today) && strtotime($today) <= strtotime($dateKT))
-            {
+            $data = $request->all();
+            $dateBD = Carbon::createFromTimestamp(strtotime($data['NGAYBD'] . $data['TDBD']));
+            $MATT = 1;
+            $dateKT = Carbon::createFromTimestamp(strtotime($data['NGAYKT'] . $data['TDKT']));
+            $today = date("Y-m-d h:i:sa");
+            if (strtotime($dateBD) <=  strtotime($today) && strtotime($today) <= strtotime($dateKT)) {
                 $MATT = 2;
             }
-                
-        $result = $dateBD->lt($dateKT);
-        $exits1 = KhuyenMai::Where('NGAYKT','>',$dateBD)->where('NGAYBD','<', $dateBD)->count();
-        $exits2 = KhuyenMai::Where('NGAYBD','<',$dateKT)->where('NGAYKT','>', $dateKT)->count();
-        
-        if ($result) 
-            {
+
+            $result = $dateBD->lt($dateKT);
+            $exits1 = KhuyenMai::Where('NGAYKT', '>', $dateBD)->where('NGAYBD', '<', $dateBD)->count();
+            $exits2 = KhuyenMai::Where('NGAYBD', '<', $dateKT)->where('NGAYKT', '>', $dateKT)->count();
+
+            if ($result) {
                 if ($exits1 > 0 || $exits2 > 0) {
-                        return redirect('admin/khuyenmai/them')->with('thatbai', 'Đã tồn tại khuyến mãi khác trong thời gian này.!');
-                }
-                else {
+                    return redirect('admin/khuyenmai/them')->with('thatbai', 'Đã tồn tại khuyến mãi khác trong thời gian này.!');
+                } else {
                     try {
                         $khuyenmai = KhuyenMai::create([
                             'TENKM' => $data['TENKM'],
@@ -81,66 +84,55 @@ class KhuyenMaiController extends Controller
                             'NGAYKT' => $dateKT,
                             'MATT' => $MATT
                         ]);
-
-                        if(isset($data['danhsach']) != null){
+                        if (isset($data['danhsach']) != null) {
                             $dskh = $data['danhsach'];
                             foreach ($dskh as $ds) {
-                                CTKhuyenMai::create([
-                                'MAKM' => $khuyenmai->MAKM,
-                                'MAKH' => $ds
-                                ]);  
-                            }            
-                        }                      
+                                $ctkhuyenmai = CTKhuyenMai::create([
+                                    'MAKM' => $khuyenmai->MAKM,
+                                    'MAKH' => $ds
+                                ]);
+                            }
+                        }
                         return redirect('admin/khuyenmai/them')->with('thongbao', 'Thêm thành công!');
                     } catch (Exception $error) {
                         return redirect('admin/khuyenmai/them')->with('thongbao', 'Thêm thất bại.!');
                     }
                 }
-            }
-            else
+            } else {
                 return redirect('admin/khuyenmai/them')->with('thatbai', 'Thời gian kết phúc phải lớn hơn thời gian bắt đầu');
+            }
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
 
     public function edit($id)
     {
-        $khoahoc = KhoaHoc::all();
-        $khuyenmai = KhuyenMai::find($id);
-        $ctkhuyenmai = CTKhuyenMai::where('MAKM', '=', $id)->get();
-        return view('admin.khuyenmai.edit', compact('khoahoc','khuyenmai','ctkhuyenmai'));
+        if (session('login')->LoaiTK == 3) {
+            $khoahoc = KhoaHoc::all();
+            $khuyenmai = KhuyenMai::find($id);
+            $ctkhuyenmai = CTKhuyenMai::where('MAKM', '=', $id)->get();
+            return view('admin.khuyenmai.edit', compact('khoahoc', 'khuyenmai', 'ctkhuyenmai'));
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $this->validate(
-            $request,
-            [
-                'TENKM' => 'unique:KhuyenMai,TENKM|min:10|max:100'
-            ],
-            [
-                'TENKM.unique' => 'Tên khuyến mãi đã tồn tại',
-                'TENKM.min' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự',
-                'TENKM.max' => 'Tên khuyến mãi phải có độ dài từ 10 đến 100 ký tự'
-            ]
-        );
-
-        $data = $request->all();   
-        $dateBD = Carbon::createFromTimestamp(strtotime($data['NGAYBD'] . $data['TDBD']));      
-        $dateKT = Carbon::createFromTimestamp(strtotime($data['NGAYKT'] . $data['TDKT']));   
-        $today = date("Y-m-d h:i:sa");
-        $MATT = 3;
-            if(strtotime($dateBD) >  strtotime($today))
-            {
+        if (session('login')->LoaiTK == 3) {
+            $data = $request->all();
+            $dateBD = Carbon::createFromTimestamp(strtotime($data['NGAYBD'] . $data['TDBD']));
+            $dateKT = Carbon::createFromTimestamp(strtotime($data['NGAYKT'] . $data['TDKT']));
+            $today = date("Y-m-d h:i:sa");
+            $MATT = 3;
+            if (strtotime($dateBD) >  strtotime($today)) {
                 $MATT = 1;
-            }    
-            else if(strtotime($dateBD) <=  strtotime($today) && strtotime($today) <= strtotime($dateKT))
-            {
+            } else if (strtotime($dateBD) <=  strtotime($today) && strtotime($today) <= strtotime($dateKT)) {
                 $MATT = 2;
             }
-        $result = $dateBD->lt($dateKT);
-        // if (KhuyenMai::where('TENKM', '=',$data['TENKM'])->count() < 1) 
-        // {
-            if($result)
-            {
+            $result = $dateBD->lt($dateKT);
+            if ($result) {
                 try {
                     $khuyenmai = KhuyenMai::find($id);
                     $khuyenmai->update(
@@ -152,34 +144,38 @@ class KhuyenMaiController extends Controller
                             'MATT' => $MATT
                         ]
                     );
-                    $dskh = $data['danhsach'];
                     CTKhuyenMai::where('MAKM', $id)->delete();
-                    if($dskh != null){
+                    if (isset($data['danhsach']) != null) {
+                        $dskh = $data['danhsach'];
+
                         foreach ($dskh as $ds) {
                             CTKhuyenMai::create([
-                            'MAKM' => $khuyenmai->MAKM,
-                            'MAKH' => $ds
-                            ]);                
+                                'MAKM' => $khuyenmai->MAKM,
+                                'MAKH' => $ds
+                            ]);
                         }
                     }
                     return redirect('admin/khuyenmai/sua/' . $id)->with('thongbao', 'Sửa thành công!');
                 } catch (Exception $error) {
                     return redirect('admin/khuyenmai/sua/' . $id)->with('thongbao', 'Sửa thất bại.!');
-                }        
+                }
+            } else {
+                return redirect('admin/khuyenmai/sua/' . $id)->with('thatbai', 'Thời gian kết phúc phải lớn hơn thời gian bắt đầu');
             }
-            else
-                return redirect('admin/khuyenmai/sua/' . $id)->with('thatbai', 'Thời gian kết phúc phải lớn hơn thời gian bắt đầu'); 
-        // }
-        // else
-        //     return redirect('admin/khuyenmai/sua/' . $id)->with('exits', 'Khuyến mãi này đã tồn tại'); 
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
 
     public function delete($id)
     {
-        $khuyenmai = KhuyenMai::find($id);
-        $khuyenmai->delete();
+        if (session('login')->LoaiTK == 3) {
+            $khuyenmai = KhuyenMai::find($id);
+            $khuyenmai->delete();
 
-        return redirect('admin/khuyenmai/')->with('thongbao', 'Xóa thành công!');
+            return redirect('admin/khuyenmai/')->with('thongbao', 'Xóa thành công!');
+        } else {
+            return view('admin.thongbao.index');
+        }
     }
-
 }
